Token güncellemeleri yapıldı

+++ Backend/app_api/controllers/ratingController.js
@@ -3,9 +3,21 @@ const Rating = require("../models/Rating");
 const addRating = async (req, res) => {
   try {
     const { bookId, rating } = req.body;
+    const userId = req.auth._id;
+    if(!rating) {
+      return res.status(400).json({
+        message: "Puan zorunludur."
+      });
+    }
+    if(rating < 1 || rating > 5) {
+      return res.status(400).json({
+        message: "Puan 1 ile 5 arasında olmalıdır."
+      });
+    }
 
     const newRating = new Rating({
       bookId,
+      user: userId,
       rating
     });
 
@@ -23,9 +35,10 @@ const updateRating = async (req, res) => {
   try {
     const { ratingId } = req.params;
     const { rating } = req.body;
+    const userId = req.auth._id;
 
     const updatedRating = await Rating.findByIdAndUpdate(
-      ratingId,
+      { _id: ratingId, user: userId },
       { rating },
       { new: true }
     );
@@ -46,8 +59,9 @@ const updateRating = async (req, res) => {
 const deleteRating = async (req, res) => {
   try {
     const { ratingId } = req.params;
+    const userId = req.auth._id;
 
-    const deletedRating = await Rating.findByIdAndDelete(ratingId);
+    const deletedRating = await Rating.findByIdAndDelete({ _id: ratingId, user: userId });
 
     if (!deletedRating) {
       return res.status(404).json({
